requirements.txt
modelscope
absl-py==2.2.2
accelerate==1.6.0
aeiou==0.0.20
aiobotocore==2.22.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aioitertools==0.12.0
aiosignal==1.3.2
alias-free-torch==0.0.6
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
appdirs==1.4.4
argbind==0.3.9
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
audiobox_aesthetics==0.0.2
audioread==3.0.1
auraloss==0.4.0
av==14.4.0
bleach==6.2.0
bokeh==3.7.3
botocore==1.37.3
braceexpand==0.1.7
Brotli==1.1.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
clean-fid==0.1.35
click==8.1.8
clip-anytorch==2.6.0
cloudpickle==3.1.1
colorcet==3.1.0
colorlog==6.9.0
configparser==7.2.0
contourpy==1.3.2
cycler==0.12.1
Cython==3.1.1
dctorch==0.1.2
decorator==4.4.2
decord==0.6.0
descript-audio-codec==1.0.0
docker-pycreds==0.4.0
docstring_parser==0.16
einops==0.7.0
einops-exts==0.0.4
ema-pytorch==0.2.3
encodec==0.1.1
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.115.12
fastcore==1.8.2
# flash-attn # NOTE: installed in cog.yaml's run section instead
ffmpeg==1.4
ffmpy==0.5.0
filelock==3.18.0
fire==0.7.0
flatten-dict==0.4.2
fonttools==4.58.0
frozenlist==1.6.0
fsspec==2025.5.0
ftfy==6.3.1
future==1.0.0
fvcore==0.1.5.post20221221
gin-config==0.5.0
gitdb==4.0.12
GitPython==3.1.44
# gradio==3.50.0 # NOTE: not used, Replicate Cog doesn't support gradio
# gradio_client==0.6.1 # NOTE: not used, Replicate Cog doesn't support gradio
groovy==0.1.2
grpcio==1.71.0
h11==0.16.0
hf_xet
h5py==3.13.0
hjson==3.1.0
holoviews==1.20.2
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.30.2
hydra-colorlog==1.2.0
hydra-core==1.3.2
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.4.9
importlib-resources==5.12.0
importlib_metadata==8.7.0
iopath==0.1.10
ipython==8.36.0
jedi==0.19.2
Jinja2==3.1.0
jmespath==1.0.1
joblib==1.5.0
jsonmerge==1.9.2
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
julius==0.2.7
k-diffusion==0.1.1
kiwisolver==1.4.8
kornia==0.8.1
kornia_rs==0.1.9
laion-clap==1.1.4
latex2mathml==3.77.0
lazy_loader==0.4
librosa==0.9.2
lightning==2.5.1.post0
lightning-utilities==0.14.3
linkify-it-py==2.0.3
llvmlite==0.43.0
local-attention==1.8.6
Markdown==3.8
markdown-it-py==3.0.0
markdown2==2.5.3
MarkupSafe==2.1.5
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
moviepy==1.0.3
mpmath==1.3.0
multidict==6.4.4
multiprocessing-logging==0.2.4
mutagen==1.47.0
narwhals==1.40.0
networkx==3.4.2
ninja==1.11.1.3
nitrous_ema==0.0.1
numba==0.60.0
numpy==1.23.5
omegaconf==2.3.0
open_clip_torch==2.32.0
openai==1.33.0
opencv-python==4.11.0.86
orjson==3.10.18
pafy==0.5.3.1
pandas==2.0.2
panel==1.7.0
param==2.2.0
parameterized==0.9.0
parso==0.8.4
pathtools==0.1.2
pedalboard==0.7.4
pexpect==4.9.0
pillow
platformdirs==4.3.8
plotly==6.1.1
pooch==1.8.2
prefigure==0.0.9
proglog==0.1.10
progressbar==2.5
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==3.19.6
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.1
pyloudnorm==0.1.1
pynndescent==0.5.13
pynvml==12.0.0
pyparsing==3.2.3
pystoi==0.4.1
pysubs2==1.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytorch-lightning==2.5.1.post0
pytorchvideo==0.1.5
pytz==2025.2
pyviz_comms==3.0.4
PyWavelets==1.4.1
PyYAML==6.0.2
randomname==0.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
resampy==0.4.3
rich==14.0.0
rpds-py==0.25.1
ruff==0.11.11
s3fs==2025.5.0
safehttpx==0.1.6
safetensors==0.5.3
scenedetect==0.6.3
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.15.3
semantic-version==2.10.0
sentencepiece==0.1.99
sentry-sdk==2.29.1
setproctitle==1.3.6
shellingham==1.5.4
shortuuid==1.0.13
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SoundFile==0.10.2
sox==1.3.0
stack-data==0.6.3
starlette==0.46.2
submitit==1.5.2
svgwrite==1.4.3
sympy==1.13.1
tabulate==0.9.0
tensorboard-data-server==0.7.2
tensordict==0.7.1
termcolor==3.1.0
threadpoolctl==3.6.0
tifffile==2025.5.10
timm==1.0.15
tokenizers==0.21.1
tomlkit==0.13.2
torch==2.6.0
torch-stoi==0.2.3
torchaudio==2.6.0
torchdiffeq==0.2.5
torchlibrosa==0.1.0
torchmetrics==0.11.4
torchsde==0.2.6
torchvision==0.21.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
trampoline==0.1.2
transformers==4.52.3
triton==3.2.0
typer==0.15.4
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
uc-micro-py==1.0.3
umap-learn==0.5.7
urllib3==2.4.0
uvicorn==0.34.2
v-diffusion-pytorch==0.0.2
vector-quantize-pytorch==1.9.14
wcwidth==0.2.13
webdataset==0.2.48
webencodings==0.5.1
Werkzeug==3.1.3
wget==3.2
wrapt==1.17.2
x-transformers==1.26.6
xyzservices==2025.4.0
yacs==0.1.8
yarl==1.20.0
zipp==3.21.0
git+https://github.com/patrick-kidger/torchcubicspline.git
altair==5.5.0
